pyproject.toml
[project]
name = "andrewaylett-pre-commit-hooks"
version = "0.1.0"
description = "Andrew's pre-commit hooks"
readme = "README.md"
authors = [
  { name = "Andrew Aylett", email = "andrew@aylett.co.uk" }
]
requires-python = ">=3.13"
dependencies = [
  "cogapp>=3.5.1",
  "ruamel.yaml>=0.17.0",
]

[dependency-groups]
dev = [
  "pytest>=7.4.0",
  "ruff>=0.12.0",
  "pre-commit>=4.2.0",
  "pyyaml>=6.0",
]

[tool.ruff.lint]
select = [
  # pycodestyle
  "E4", "E7", "E9",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # Ruff
  "RUF",
  # isort
  "I",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[project.scripts]
pre-commit-cog = "andrewaylett_pre_commit_hooks.cog:main"
pre-commit-init-hooks = "andrewaylett_pre_commit_hooks.init_hooks:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
